Finally after 3 days of perseverance, i designed HashSet using Doubly Linked List and Array in JavaScript
I chained my Doubly Linked Lists with Array to handle collisions, and used Modulus as my hashing function!

yesterday all testcases were passing when i was setting the size of Array as 10, but when i was increasing its size to 100 or more then test cases on leetcode were breaking, i dry run a lot and confused, finally i found that the issue was with deleting the end node from doubly linked list, i was not setting its next node as null, and previous node was keep pointing to deleted node.
feeling happy